fixes #30, migrate Versions page, add it to the dropdown
