+++ website/src/css/versions.scss
@@ -1,0 +1,38 @@
+@import "shared";
+
+.versions-page {
+  max-width: 960px;
+  padding: 28px;
+
+  h1 {
+    font-size: 3rem;
+  }
+
+  h2 {
+    font-size: 2rem;
+  }
+
+  p a,
+  td a {
+    @extend %link-style;
+
+    code {
+      background: none;
+      white-space: nowrap;
+    }
+  }
+
+  table th,
+  table td {
+    min-width: 100px;
+    font-size: 15px;
+    padding: 8px 20px;
+  }
+}
+
+html[data-theme="dark"] .versions-page {
+  p a,
+  td a {
+    @extend %link-style-dark;
+  }
+}
